requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
facepy==1.0.9
hupper==1.1
idna==2.6
jsonschema==2.6.0
PasteDeploy==1.5.2
plaster==1.0
plaster-pastedeploy==0.5
pydot==1.2.4
PyJWT==1.6.1
pyparsing==2.2.0
pyramid==1.9.1
pyramid-mailer==0.15.1
pyramid-rest-framework==0.0.6
pytz==2018.3
repoze.lru==0.7
repoze.sendmail==4.4.1
requests==2.18.4
Rest-Framework-Auth-Toolkit==0.7
rest-framework-auth0==0.4.6
rest-framework-borderkeeper==0.0.2
rest-framework-dyn-serializer==1.3.1
rfc3987==1.3.7
six==1.11.0
SQLAHelper==1.0
SQLAlchemy==1.2.6
sqlalchemy-schemadisplay==1.3
strict-rfc3339==0.7
transaction==2.2.1
translationstring==1.3
urllib3==1.22
venusian==1.1.0
WebOb==1.8.0rc1
zope.deprecation==4.3.0
zope.interface==4.4.3
zope.sqlalchemy==1.0
